+++ src/main/java/net/jodah/expiringmap/NamedThreadFactory.java
@@ -1,43 +1,39 @@
+/*
+ * Copyright 2014 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package net.jodah.expiringmap;
 
 import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.ThreadPoolExecutor;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
- * Decorates a thread factory with named threads.
- * 
- * @author Jonathan Halterman
+ * Named thread factory.
  */
 class NamedThreadFactory implements ThreadFactory {
-  private final String name;
-  private final ThreadFactory threadFactory;
+  private final AtomicInteger threadNumber = new AtomicInteger(1);
+  private final String nameFormat;
 
   /**
-   * Creates a new NamedThreadFactory object.
-   * 
-   * @param threadFactory Factory to decorate
-   * @param name Name
+   * Creates a thread factory that names threads according to the {@code nameFormat} by supplying a
+   * single argument to the format representing the thread number.
    */
-  NamedThreadFactory(ThreadFactory threadFactory, String name) {
-    this.threadFactory = threadFactory;
-    this.name = name + " ";
-  }
-
-  /**
-   * Decorates the given thread pool executor with a named thread factory.
-   * 
-   * @param executor Executor to decorate.
-   * @return The decorated executor
-   */
-  static ThreadPoolExecutor decorate(ThreadPoolExecutor executor, String pName) {
-    executor.setThreadFactory(new NamedThreadFactory(executor.getThreadFactory(), pName));
-    return executor;
+  public NamedThreadFactory(String nameFormat) {
+    this.nameFormat = nameFormat;
   }
 
   @Override
-  public Thread newThread(Runnable runnable) {
-    Thread thread = threadFactory.newThread(runnable);
-    thread.setName(name + thread.getName());
-    return thread;
+  public Thread newThread(Runnable r) {
+    return new Thread(r, String.format(nameFormat, threadNumber.getAndIncrement()));
   }
 }